used react router to switch between pages
