+++ src/components/app/App.js
@@ -1,46 +1,24 @@
-import { useState } from "react";
-import AppHeader from "../appHeader/AppHeader";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import AppBanner from "../appBanner/AppBanner";
-import ComicsList from "../comicsList/ComicsList";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
+import AppHeader from "../appHeader/AppHeader";
+import {MainPage, ComicsPage} from "../pages";
 
 const App = () => {
-    const [selectedChar, setSelectedChar] = useState(null);
-
-    const onCharSelected = (id) => {
-        setSelectedChar(id === selectedChar ? null : id)
-    }
-
     return (
-        <div className="app">
-            <AppHeader/>
-            {/* Characters */}
-            {/* <main>
-                <ErrorBoundary>
-                    <RandomChar/>
-                </ErrorBoundary>
-                <div className="char__content">
-                    <ErrorBoundary>
-                        <CharList onCharSelected={onCharSelected}/>
-                    </ErrorBoundary>
-                    <ErrorBoundary>
-                        <CharInfo charId={selectedChar}/>
-                    </ErrorBoundary>
-                </div>
-                <img className="bg-decoration" src={decoration} alt="vision"/>
-            </main> */}
-
-            {/* Comics */}
-            <main>
-                <AppBanner/>
-                <ErrorBoundary>
-                    <ComicsList/>
-                </ErrorBoundary>
-            </main>
-        </div>
+        <Router>
+            <div className="app">
+                <AppHeader/>
+                <main>
+                    <Routes>
+                        {/* Characters */}
+                        <Route path="/" element={<MainPage/>}></Route>
+                        {/* Comics */}
+                        <Route path="/comics" element={<ComicsPage/>}></Route>
+                    </Routes>
+                </main>
+            </div>
+        </Router>
     )
-    
 }
 
 export default App;
