+++ src/components/appHeader/appHeader.scss
@@ -25,6 +25,9 @@
                 a:hover {
                     color: $main-color;
                 }
+                .active {
+                    color: $main-color;
+                }
             }
         }
     }
