+++ src/components/pages/ComicsPage.js
@@ -1,0 +1,17 @@
+import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+import AppBanner from "../appBanner/AppBanner";
+import ComicsList from "../comicsList/ComicsList";
+
+
+const ComicsPage = () => {
+    return (
+        <>
+            <AppBanner/>
+            <ErrorBoundary>
+                <ComicsList/>
+            </ErrorBoundary>
+        </>
+    )
+}
+
+export default ComicsPage;
